requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
openai==1.77.0
python-dotenv==0.19.2
celery==5.3.6
redis==5.0.1
opencv-python
